Put all classes into a java package
There is a lack of organisation where all the classes are in the
same folder.

Let's put all the classes into a java package.
