+++ src/main/java/tako/Parser.java
@@ -1,4 +1,16 @@
-import java.io.IOException;
+package tako;
+
+import tako.command.Command;
+import tako.command.AddCommand;
+import tako.command.DeleteCommand;
+import tako.command.ExitCommand;
+import tako.command.ListCommand;
+import tako.command.MarkCommand;
+
+import tako.task.Deadline;
+import tako.task.Event;
+import tako.task.Todo;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
