+++ src/main/java/tako/Tako.java
@@ -1,3 +1,7 @@
+package tako;
+
+import tako.command.Command;
+
 import java.io.IOException;
 
 /**
