+++ src/main/java/tako/TakoException.java
@@ -1,3 +1,5 @@
+package tako;
+
 public class TakoException extends Exception{
     public TakoException(String message) {
         super(message);
