+++ src/main/java/tako/command/Command.java
@@ -1,3 +1,10 @@
+package tako.command;
+
+import tako.Storage;
+import tako.TakoException;
+import tako.TaskList;
+import tako.Ui;
+
 import java.io.IOException;
 
 public abstract class Command {
