+++ src/main/java/tako/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package tako.command;
+
+import tako.Storage;
+import tako.TaskList;
+import tako.Ui;
+
 public class ExitCommand extends Command{
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
