+++ src/main/java/tako/task/Event.java
@@ -1,3 +1,5 @@
+package tako.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
