+++ src/main/java/tako/task/Task.java
@@ -1,3 +1,5 @@
+package tako.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
